+++ cmp/filter/FilterChooserModel.js
@@ -4,7 +4,7 @@
  *
  * Copyright © 2021 Extremely Heavy Industries Inc.
  */
-import {HoistModel, managed, PersistenceProvider, XH} from '@xh/hoist/core';
+import {HoistModel, managed, PersistenceProvider, XH, TaskObserver} from '@xh/hoist/core';
 import {FieldFilter, parseFilter, combineValueFilters} from '@xh/hoist/data';
 import {action, observable, makeObservable} from '@xh/hoist/mobx';
 import {wait} from '@xh/hoist/promise';
@@ -56,6 +56,9 @@ export class FilterChooserModel extends HoistModel {
     @managed provider;
     persistValue = false;
     persistFavorites = false;
+
+    /** @member {TaskObserver} - tracks execution of filtering operation on bound object.*/
+    @managed filterTask = TaskObserver.trackAll();
 
     // Implementation fields for Control
     @observable.ref selectOptions;
@@ -154,7 +157,11 @@ export class FilterChooserModel extends HoistModel {
         if (bind) {
             this.addReaction({
                 track: () => this.value,
-                run: (filter) => bind.setFilter(filter),
+                run: (filter) => {
+                    wait()
+                        .then(() => bind.setFilter(filter))
+                        .linkTo(this.filterTask);
+                },
                 fireImmediately: true
             });