+++ CHANGELOG.md
@@ -10,6 +10,8 @@
     * New `Column.filterable` config enables a customized header menu with filtering options. The new
       control offers two tabs - a "Values" tab for an enumerated "set-type" filter and a "Custom" tab
       to support more complex queries with multiple clauses.
+* New `TaskObserver` replaces existing `PendingTaskModel`, providing improved support for joining
+  and masking multiple asynchronous tasks.
 * New `Cube` aggregators `ChildCountAggregator` and `LeafCountAggregator`.
 
 ### 💥 Breaking Changes
@@ -18,6 +20,10 @@
   `FilterChooserModel.valueSource` and `FilterChooserModel.bind` respectively. Furthermore, both
   configs now support either a `Store` or a cube `View`. This is to provide a common API with the
   new `GridFilterModel` filtering described above.
+* `PendingTaskModel` has been replaced by the new `TaskObserver` in `@xh/hoist/core`.
+* The `model` prop on `LoadingIndicator` and `Mask` has been replaced with `bind`.  Provide one or
+  more `TaskObserver`s to this prop.
+
 
 [Commit Log](https://github.com/xh/hoist-react/compare/v41.2.0...develop)